self-healing-agent/requirements.txt
fastapi
uvicorn
langgraph
langchain-openai
python-dotenv
requests